Moved ModuleUtils into mill.scalalib.internal
It belongs to scalalib and we want to avoid split packages.
